fix: restore live Electron daemon bridge boot
